refactor(cas): split the crate into domain, webauthn and http modules
Mechanical move, no logic change: `accounts/` (types, display_name,
repository), `webauthn/` (ceremonies, passkeys, registration and the
relying-party builder), and `http/` (router assembly, ApiError, extractors,
health, the /api/webauthn handlers). Domain modules never import axum; the
HTTP layer owns the domain-error to status mappings next to its handlers.

With the transport in the library the binary has no tests left, so the
`test-support` feature and the crate's dev-dependency on itself go away;
`testing` is a plain cfg(test) module and webauthn-authenticator-rs is an
ordinary dev-dependency again. AGENTS.md documents the layout.
