+++ apps/cas/src/accounts/display_name.rs
@@ -1,16 +1,9 @@
-//! Accounts — the identity root of CAS.
-//!
-//! An account id is the OIDC `sub`. Guests are ordinary rows
-//! (`type = AccountType::Guest`) so sessions, `/me` and userinfo have a single
-//! code path for guests and full accounts; a guest that registers a passkey is
-//! upgraded in place, keeping its `sub`. See `docs/PLAN.md`.
+//! The display name newtype: how a user-facing name is sanitised, validated
+//! and carried across the database boundary.
 
 use nutype::nutype;
-use sqlx::PgPool;
-use time::OffsetDateTime;
 use unicode_general_category::GeneralCategory;
 use unicode_normalization::UnicodeNormalization;
-use uuid::Uuid;
 
 /// Upper bound on a display name, in characters. Long enough for a real name,
 /// short enough to keep the sign-in UI and the authenticator prompt readable.
@@ -199,154 +192,6 @@ impl sqlx::Encode<'_, sqlx::Postgres> for DisplayName {
     }
 }
 
-/// Whether an account has credentials of its own.
-///
-/// Maps to the Postgres `account_type` enum.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, sqlx::Type)]
-#[sqlx(type_name = "account_type", rename_all = "lowercase")]
-pub enum AccountType {
-    /// Temporary identity created without any user interaction.
-    Guest,
-    /// Account with at least one credential (a passkey).
-    Full,
-}
-
-/// A row of `accounts`.
-#[derive(Debug, Clone, PartialEq, Eq, sqlx::FromRow)]
-pub struct Account {
-    /// Stable account id, exposed as the OIDC `sub`.
-    pub id: Uuid,
-    /// Non-unique: CAS has no username in v1. Always a valid [`DisplayName`]:
-    /// validated when the row is decoded, so a row that was written or altered
-    /// outside CAS and does not pass surfaces as a decode error instead of
-    /// reaching a UI.
-    pub display_name: DisplayName,
-    pub r#type: AccountType,
-    /// Optional and unverified in v1; the future recovery anchor.
-    pub email: Option<String>,
-    pub created_at: OffsetDateTime,
-    /// Refreshed on activity; drives guest GC.
-    pub last_seen_at: OffsetDateTime,
-}
-
-/// The values a caller supplies when creating an account. `created_at` and
-/// `last_seen_at` are assigned by the database.
-#[derive(Debug, Clone)]
-pub struct NewAccount {
-    /// Generated by CAS, not by the database: passkey registration hands the id
-    /// to the authenticator as the WebAuthn user handle before the row exists.
-    /// The handle is baked into the credential and can never change, and
-    /// discoverable login (#666) gets it back as the account to sign in.
-    /// Defaults to a fresh v4 uuid; [`NewAccount::with_id`] pins a chosen one.
-    pub id: Uuid,
-    pub display_name: DisplayName,
-    pub r#type: AccountType,
-    pub email: Option<String>,
-}
-
-impl NewAccount {
-    /// A guest: no credentials, no email.
-    pub fn guest(display_name: DisplayName) -> Self {
-        Self {
-            id: Uuid::new_v4(),
-            display_name,
-            r#type: AccountType::Guest,
-            email: None,
-        }
-    }
-
-    /// A full account. Email stays optional — registration only asks for a
-    /// display name.
-    pub fn full(display_name: DisplayName) -> Self {
-        Self {
-            id: Uuid::new_v4(),
-            display_name,
-            r#type: AccountType::Full,
-            email: None,
-        }
-    }
-
-    /// Uses an id chosen by the caller instead of the generated one. Passkey
-    /// registration needs this: the id travels to the authenticator when the
-    /// ceremony starts, the row is written when it finishes.
-    #[must_use]
-    pub fn with_id(mut self, id: Uuid) -> Self {
-        self.id = id;
-        self
-    }
-
-    #[must_use]
-    pub fn with_email(mut self, email: impl Into<String>) -> Self {
-        self.email = Some(email.into());
-        self
-    }
-}
-
-/// Inserts an account with any executor — a pool, or the transaction that
-/// registration uses to write an account and its first passkey together.
-pub(crate) async fn insert<'e, E>(executor: E, account: NewAccount) -> Result<Account, sqlx::Error>
-where
-    E: sqlx::PgExecutor<'e>,
-{
-    sqlx::query_as!(
-        Account,
-        r#"INSERT INTO accounts (id, display_name, type, email)
-           VALUES ($1, $2, $3, $4)
-           RETURNING
-               id,
-               display_name AS "display_name: DisplayName",
-               type AS "type: AccountType",
-               email,
-               created_at,
-               last_seen_at"#,
-        account.id,
-        // `as _`: the macro infers `&str` for a text parameter, so the cast is
-        // what makes it use the `Encode` impl for the newtype instead.
-        account.display_name as _,
-        account.r#type as AccountType,
-        account.email,
-    )
-    .fetch_one(executor)
-    .await
-}
-
-/// Data access for `accounts`.
-#[derive(Debug, Clone)]
-pub struct AccountRepository {
-    pool: PgPool,
-}
-
-impl AccountRepository {
-    pub fn new(pool: PgPool) -> Self {
-        Self { pool }
-    }
-
-    /// Inserts an account and returns the stored row, including the
-    /// database-assigned timestamps.
-    pub async fn create(&self, account: NewAccount) -> Result<Account, sqlx::Error> {
-        insert(&self.pool, account).await
-    }
-
-    /// Looks an account up by id. `Ok(None)` means no such account.
-    pub async fn get(&self, id: Uuid) -> Result<Option<Account>, sqlx::Error> {
-        sqlx::query_as!(
-            Account,
-            r#"SELECT
-                   id,
-                   display_name AS "display_name: DisplayName",
-                   type AS "type: AccountType",
-                   email,
-                   created_at,
-                   last_seen_at
-               FROM accounts
-               WHERE id = $1"#,
-            id,
-        )
-        .fetch_optional(&self.pool)
-        .await
-    }
-}
-
 #[cfg(test)]
 mod display_name_tests {
     use super::*;
@@ -482,114 +327,5 @@ mod display_name_tests {
 
         let ok: DisplayName = serde_json::from_str("\"  Ada  \"").unwrap();
         assert_eq!(ok.as_ref(), "Ada");
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::testing::display_name;
-
-    #[sqlx::test]
-    async fn creates_a_full_account_with_an_email(pool: PgPool) {
-        let repository = AccountRepository::new(pool);
-
-        let account = repository
-            .create(NewAccount::full(display_name("Ada")).with_email("ada@example.com"))
-            .await
-            .unwrap();
-
-        assert_eq!(account.display_name.as_ref(), "Ada");
-        assert_eq!(account.r#type, AccountType::Full);
-        assert_eq!(account.email.as_deref(), Some("ada@example.com"));
-        // A fresh account has never been seen after its creation.
-        assert_eq!(account.last_seen_at, account.created_at);
-    }
-
-    #[sqlx::test]
-    async fn creates_a_guest_without_an_email(pool: PgPool) {
-        let repository = AccountRepository::new(pool);
-
-        let account = repository
-            .create(NewAccount::guest(display_name("Guest")))
-            .await
-            .unwrap();
-
-        assert_eq!(account.r#type, AccountType::Guest);
-        assert_eq!(account.email, None);
-    }
-
-    #[sqlx::test]
-    async fn get_returns_the_created_account(pool: PgPool) {
-        let repository = AccountRepository::new(pool);
-        let created = repository
-            .create(NewAccount::full(display_name("Grace")))
-            .await
-            .unwrap();
-
-        let found = repository.get(created.id).await.unwrap();
-
-        assert_eq!(found, Some(created));
-    }
-
-    #[sqlx::test]
-    async fn get_returns_none_for_an_unknown_id(pool: PgPool) {
-        let repository = AccountRepository::new(pool);
-
-        let found = repository.get(Uuid::new_v4()).await.unwrap();
-
-        assert_eq!(found, None);
-    }
-
-    #[sqlx::test]
-    async fn creates_an_account_with_a_caller_chosen_id(pool: PgPool) {
-        let repository = AccountRepository::new(pool);
-        let id = Uuid::new_v4();
-
-        let account = repository
-            .create(NewAccount::full(display_name("Ada")).with_id(id))
-            .await
-            .unwrap();
-
-        assert_eq!(account.id, id);
-    }
-
-    /// The table has no CHECK, so nothing stops a row written outside CAS from
-    /// holding a name with a bidi override in it. The reader is the guarantee:
-    /// such a row fails to decode instead of reaching a UI.
-    #[sqlx::test]
-    async fn get_fails_to_decode_an_invalid_display_name(pool: PgPool) {
-        let id = Uuid::new_v4();
-        // Unchecked query: see docs/TESTS.md.
-        sqlx::query("INSERT INTO accounts (id, display_name, type) VALUES ($1, $2, 'full')")
-            .bind(id)
-            .bind("\u{202e}adA")
-            .execute(&pool)
-            .await
-            .unwrap();
-
-        let error = AccountRepository::new(pool).get(id).await.unwrap_err();
-
-        assert!(
-            matches!(error, sqlx::Error::ColumnDecode { .. }),
-            "expected a column decode error, got {error:?}"
-        );
-    }
-
-    /// v1 has no username: two accounts may share a display name.
-    #[sqlx::test]
-    async fn display_names_are_not_unique(pool: PgPool) {
-        let repository = AccountRepository::new(pool);
-
-        let first = repository
-            .create(NewAccount::full(display_name("Ada")))
-            .await
-            .unwrap();
-        let second = repository
-            .create(NewAccount::full(display_name("Ada")))
-            .await
-            .unwrap();
-
-        assert_ne!(first.id, second.id);
     }
 }
