+++ apps/cas/src/accounts/mod.rs
@@ -1,0 +1,99 @@
+//! Accounts — the identity root of CAS.
+//!
+//! An account id is the OIDC `sub`. Guests are ordinary rows
+//! (`type = AccountType::Guest`) so sessions, `/me` and userinfo have a single
+//! code path for guests and full accounts; a guest that registers a passkey is
+//! upgraded in place, keeping its `sub`. See `docs/PLAN.md`.
+
+mod display_name;
+mod repository;
+
+use time::OffsetDateTime;
+use uuid::Uuid;
+
+pub use display_name::{DisplayName, DisplayNameError, MAX_DISPLAY_NAME_LENGTH};
+pub use repository::AccountRepository;
+pub(crate) use repository::insert;
+
+/// Whether an account has credentials of its own.
+///
+/// Maps to the Postgres `account_type` enum.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, sqlx::Type)]
+#[sqlx(type_name = "account_type", rename_all = "lowercase")]
+pub enum AccountType {
+    /// Temporary identity created without any user interaction.
+    Guest,
+    /// Account with at least one credential (a passkey).
+    Full,
+}
+
+/// A row of `accounts`.
+#[derive(Debug, Clone, PartialEq, Eq, sqlx::FromRow)]
+pub struct Account {
+    /// Stable account id, exposed as the OIDC `sub`.
+    pub id: Uuid,
+    /// Non-unique: CAS has no username in v1. Always a valid [`DisplayName`]:
+    /// validated when the row is decoded, so a row that was written or altered
+    /// outside CAS and does not pass surfaces as a decode error instead of
+    /// reaching a UI.
+    pub display_name: DisplayName,
+    pub r#type: AccountType,
+    /// Optional and unverified in v1; the future recovery anchor.
+    pub email: Option<String>,
+    pub created_at: OffsetDateTime,
+    /// Refreshed on activity; drives guest GC.
+    pub last_seen_at: OffsetDateTime,
+}
+
+/// The values a caller supplies when creating an account. `created_at` and
+/// `last_seen_at` are assigned by the database.
+#[derive(Debug, Clone)]
+pub struct NewAccount {
+    /// Generated by CAS, not by the database: passkey registration hands the id
+    /// to the authenticator as the WebAuthn user handle before the row exists.
+    /// The handle is baked into the credential and can never change, and
+    /// discoverable login (#666) gets it back as the account to sign in.
+    /// Defaults to a fresh v4 uuid; [`NewAccount::with_id`] pins a chosen one.
+    pub id: Uuid,
+    pub display_name: DisplayName,
+    pub r#type: AccountType,
+    pub email: Option<String>,
+}
+
+impl NewAccount {
+    /// A guest: no credentials, no email.
+    pub fn guest(display_name: DisplayName) -> Self {
+        Self {
+            id: Uuid::new_v4(),
+            display_name,
+            r#type: AccountType::Guest,
+            email: None,
+        }
+    }
+
+    /// A full account. Email stays optional — registration only asks for a
+    /// display name.
+    pub fn full(display_name: DisplayName) -> Self {
+        Self {
+            id: Uuid::new_v4(),
+            display_name,
+            r#type: AccountType::Full,
+            email: None,
+        }
+    }
+
+    /// Uses an id chosen by the caller instead of the generated one. Passkey
+    /// registration needs this: the id travels to the authenticator when the
+    /// ceremony starts, the row is written when it finishes.
+    #[must_use]
+    pub fn with_id(mut self, id: Uuid) -> Self {
+        self.id = id;
+        self
+    }
+
+    #[must_use]
+    pub fn with_email(mut self, email: impl Into<String>) -> Self {
+        self.email = Some(email.into());
+        self
+    }
+}
