+++ apps/cas/src/http/health.rs
