+++ apps/cas/src/lib.rs
@@ -1,10 +1,9 @@
-// Library part of the crate: the domain and the modules shared between the
-// `cas` server binary and the `cas-migrate` binary.
+// Library part of the crate: the domain, the HTTP transport, and the modules
+// shared between the `cas` server binary and the `cas-migrate` binary.
 pub mod accounts;
-pub mod ceremonies;
 pub mod config;
+pub mod http;
 pub mod migrations;
-pub mod passkeys;
-pub mod registration;
-#[cfg(feature = "test-support")]
-pub mod testing;
+#[cfg(test)]
+pub(crate) mod testing;
+pub mod webauthn;
